stages/rpm: optional to disable to run dracut
Add the ability to prevent dracut from being run via the post
install scripts by installing a symlink to /dev/null in the
/etc/kernel/install.d directory which takes precedent over the
system maintained /usr/lib/kernel/install.d/ which actually
contains the scripts that run dracut.
After the run is completed the symlinks are removed and the
original files thus unmasked.